+++ src/main/java/app/onepass/organizer/entities/EventTagEntity.java
@@ -25,9 +25,9 @@ public class EventTagEntity implements BaseEntity<EventTagMessage, EventTagEntit
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
-	private long eventId;
-	private long tagId;
+	private int id;
+	private int eventId;
+	private int tagId;
 
 	@Override
 	public EventTagMessage parseEntity() {
